resolver: inline small tls module
